es-visitor: add support for BAI
NEW: Add support for BAI author queries, which are generated when an
author keyword is detected and either the value of the node matches the
BAI pattern which queries only the BAI field, or the value contains a
dot and no whitespace (i.e. author name), which queries both authors
name and BAI fields.

Signed-off-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>