es-visitor: date queries use ES date rounding
This is needed for supporting range queries on dates, i.e. rounding
them up or down according to the ES range operator.

For example, without this, a query like 'date > 2010-11', would return
documents with date '2010-11-15', due to the date value of the query
being interpreted by ES as '2010-11-01 01:00:00'.

By using the suffixes for rounding up or down, the date value of the
query is interpreted as '2010-11-30T23:59:59.999', thus not returning
the document with date '2010-11-15', as the user would expect. See:
https://www.elastic.co/guide/en/elasticsearch/reference/6.1/query-dsl-range-query.html#_date_math_and_rounding

Signed-off-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>