Add journal nested queries support
* es-visitor: add support for journal queries on journal title, journal
volume and artid/page start.

* restructuring-visitor: implement `journal and vol` query support and
drop the `vol` part of `journal and not vol` queries.

* setup: depend on `inspire-schemas`, since we require `convert_old_publication_info_to_new`.
For that reason, `inspire-utils` is dropped as a dependency, in place of `inspire-schemas`.

* `convert_old_publication_info_to_new` of `inspire-schemas` is used
for supporting journal keyword queries.

* Use `convert_old_publication_info_to_new` of `inspire-schemas` to
preprocess and validate the searched values, i.e. the journal_title,
journal_volume and artid/page_start.

* utils: add generic functions for generating nested, match
and bool queries.

* Ensure journal queries are the same for either Value, ExactValue and
PartialValue.

Sem-Ver: new feature

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>
Co-authored-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>