es-visitor: improve title search
Add support for searching on titles (usually) using terminology, like
"g-2" and "SU(2)" - which requires from our index, tokens to be
whitespace tokenized ("search" field variation of titles).

We have to filter the tokens which we query against this search field.
The reason for this, is supporting queries that try to match on parts
of the whitespace tokens.
For example, for a query like "t n-body two" to be able to match the
record "N body ... two-body", we shouldn't query the search field on
with the "two" because it's not going to be there. Thus, we have to
differentiate among terminology tokens.

Signed-off-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>