es-visitor: add support for type-code queries
The type-code query requires a special way of hanlding it, as in
depending on the value of the type-code query, we query another field
in accordance to the `TYPECODE_VALUE_TO_FIELD_AND_VALUE_PAIRS_MAPPING`.
Even if the value of the query is an exact, or partial value then it is
handled in the same way as a simple value (closes #79).

Signed-off-by: Chris Aslanoglou <chris.aslanoglou@gmail.com>